Remove converters TODO -- not useful
The listed data types on the CKAN page does not include `numeric`, which we are getting
from CKAN now. So, who knows if we would ever get the types listed there.

The error messages for a missing type are hopefully helpful enough to make fixing this
easy in the future if we get new types of data from CKAN.