Parameterize CKAN queries
Plus add some safety

Error handling slight refactor:

Steps that produce errors now warn to the console for debugging, since they can
print useful context there that may not be user-friendly.

Error messages are changed to translation keys.

A basic translations interpolation implementation is included in this commit...
but it might be time to look to adding a real translations tool for that.