Optimize clustering
- throttle filter updates
- disable prune cluster insertion sort

The problem with PruneCluster's insertionSort is that it chooses it based on number of markers added as a percentage of the total
(set to 20%).

The problem is that 20% of the total becomes a large number as the total becomes large, and the cost of the insertion sort grows
exponentially.

For now, it is disabled entirely. An ideal solution would find a more robust way of figuring out when the insertion sort is likely
to be faster.